Rename package 'nft.activity' to 'nft.events'
