+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/nft/collection/NftCollectionFragment.kt
@@ -14,8 +14,8 @@ import androidx.viewpager2.adapter.FragmentStateAdapter
 import io.horizontalsystems.bankwallet.R
 import io.horizontalsystems.bankwallet.core.BaseFragment
 import io.horizontalsystems.bankwallet.databinding.FragmentCollectionBinding
-import io.horizontalsystems.bankwallet.modules.nft.collection.activity.NftCollectionEventsFragment
 import io.horizontalsystems.bankwallet.modules.nft.collection.assets.NftCollectionAssetsFragment
+import io.horizontalsystems.bankwallet.modules.nft.collection.events.NftCollectionEventsFragment
 import io.horizontalsystems.bankwallet.modules.nft.collection.overview.NftCollectionOverviewFragment
 import io.horizontalsystems.bankwallet.ui.compose.ComposeAppTheme
 import io.horizontalsystems.bankwallet.ui.compose.components.TabItem
