+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/nft/collection/events/NftCollectionEventsViewModel.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.bankwallet.modules.nft.collection.activity
+package io.horizontalsystems.bankwallet.modules.nft.collection.events
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
